+++ tiled/_tests/test_access_control.py
@@ -2,16 +2,71 @@ import json
 
 import numpy
 import pytest
-from starlette.status import HTTP_403_FORBIDDEN
+from fastapi import HTTPException
+from starlette.status import HTTP_403_FORBIDDEN, HTTP_404_NOT_FOUND
 
+from ..access_policies import (
+    ALL_SCOPES,
+    PUBLIC_SCOPES,
+    SimpleAccessPolicy,
+    SpecialUsers,
+)
 from ..adapters.array import ArrayAdapter
 from ..adapters.mapping import MapAdapter
 from ..client import Context, from_context
+from ..client.utils import ClientError
 from ..server.app import build_app_from_config
+from ..server.core import NoEntry
 from .utils import enter_username_password, fail_with_status_code
 
 arr = numpy.ones((5, 5))
+arr_zeros = numpy.zeros((5, 5))
 arr_ad = ArrayAdapter.from_array(arr)
+
+
+class EntryBasedAccessPolicy(SimpleAccessPolicy):
+    """
+    This example access policy demonstrates how the metadata on some nested child node
+    can be efficiently consulted and incorporated in logic that determines access scopes.
+    In this test example, the metadata on the node quite literally lists some scopes that
+    it should not allow. In realistic examples it could be incorporated in site-specific logic.
+    """
+
+    async def allowed_scopes(self, node, principal, path_parts):
+        # If this is being called, filter_access has let us get this far.
+        if principal is SpecialUsers.public:
+            allowed = PUBLIC_SCOPES
+        elif principal.type == "service":
+            allowed = self.scopes
+        else:
+            allowed = self.scopes
+
+        if self._get_id(principal) in self.admins:
+            allowed = ALL_SCOPES
+        else:
+            # Allowed scopes will be filtered based on some metadata of the target entry
+            try:
+                for i, segment in enumerate(path_parts):
+                    if hasattr(node, "lookup_adapter"):
+                        node = await node.lookup_adapter(path_parts[i:])
+                        if node is None:
+                            raise NoEntry(path_parts)
+                        break
+                    else:
+                        try:
+                            node = node[segment]
+                        except (KeyError, TypeError):
+                            raise NoEntry(path_parts)
+            except NoEntry:
+                raise HTTPException(
+                    status_code=HTTP_404_NOT_FOUND,
+                    detail=f"No such entry: {path_parts}",
+                )
+            remove_scope = node.metadata().get("remove_scope", None)
+            if remove_scope in allowed:
+                allowed = allowed.copy()
+                allowed.remove(remove_scope)
+        return allowed
 
 
 def tree_a(access_policy=None):
@@ -48,10 +103,10 @@ def context(tmpdir_module):
         "access_control": {
             "access_policy": "tiled.access_policies:SimpleAccessPolicy",
             "args": {
-                "access_lists": {"alice": ["a", "c", "d", "e"]},
+                "access_lists": {"alice": ["a", "c", "d", "e", "g", "h"]},
                 "provider": "toy",
                 "admins": ["admin"],
-                "public": ["f"],
+                "public": ["f", "g"],
             },
         },
         "trees": [
@@ -128,6 +183,34 @@ def context(tmpdir_module):
                 },
             },
             {"tree": ArrayAdapter.from_array(arr), "path": "/f"},
+            {
+                "tree": "tiled.catalog:in_memory",
+                "args": {"writable_storage": tmpdir_module / "g"},
+                "path": "/g",
+                "access_control": {
+                    "access_policy": "tiled.access_policies:SimpleAccessPolicy",
+                    "args": {
+                        "provider": "toy",
+                        "key": "project",
+                        "access_lists": {"alice": ["projectA"], "bob": ["projectB"]},
+                        "admins": ["admin"],
+                        "public": ["projectC"],
+                    },
+                },
+            },
+            {
+                "tree": "tiled.catalog:in_memory",
+                "args": {"writable_storage": tmpdir_module / "h"},
+                "path": "/h",
+                "access_control": {
+                    "access_policy": "tiled._tests.test_access_control:EntryBasedAccessPolicy",
+                    "args": {
+                        "provider": "toy",
+                        "access_lists": {"alice": ["x", "y"]},
+                        "admins": ["admin"],
+                    },
+                },
+            },
         ],
     }
     app = build_app_from_config(config)
@@ -138,7 +221,19 @@ def context(tmpdir_module):
                 admin_client[k].write_array(arr, key="A1")
                 admin_client[k].write_array(arr, key="A2")
                 admin_client[k].write_array(arr, key="x")
+            for k, v in {"A3": "projectA", "A4": "projectB", "r": "projectC"}.items():
+                admin_client["g"].write_array(arr, key=k, metadata={"project": v})
+            for k, v in {"x": "write:data", "y": None}.items():
+                admin_client["h"].write_array(arr, key=k, metadata={"remove_scope": v})
         yield context
+
+
+def test_entry_based_scopes(context, enter_username_password):
+    with enter_username_password("alice", "secret1"):
+        alice_client = from_context(context, username="alice")
+    with pytest.raises(ClientError, match="Not enough permissions"):
+        alice_client["h"]["x"].write(arr_zeros)
+    alice_client["h"]["y"].write(arr_zeros)
 
 
 def test_top_level_access_control(context, enter_username_password):
@@ -148,9 +243,15 @@ def test_top_level_access_control(context, enter_username_password):
     assert "A2" in alice_client["a"]
     assert "A1" not in alice_client["a"]
     assert "b" not in alice_client
+    assert "g" in alice_client
+    assert "A3" in alice_client["g"]
+    assert "A4" not in alice_client["g"]
     alice_client["a"]["A2"]
+    alice_client["g"]["A3"]
     with pytest.raises(KeyError):
         alice_client["b"]
+    with pytest.raises(KeyError):
+        alice_client["g"]["A4"]
 
     with enter_username_password("bob", "secret2"):
         bob_client = from_context(context, username="bob")
@@ -159,6 +260,8 @@ def test_top_level_access_control(context, enter_username_password):
         bob_client["a"]
     with pytest.raises(KeyError):
         bob_client["b"]
+    with pytest.raises(KeyError):
+        bob_client["g"]["A3"]
     alice_client.logout()
 
     # Make sure clearing default identity works without raising an error.
@@ -177,6 +280,7 @@ def test_access_control_with_api_key_auth(context, enter_username_password):
         context.api_key = key_info["secret"]
         client = from_context(context)
         client["a"]["A2"]
+        client["g"]["A3"]
     finally:
         # Clean up Context, which is a module-scopae fixture shared with other tests.
         context.api_key = None
@@ -194,7 +298,11 @@ def test_node_export(enter_username_password, context, buffer):
     assert "A2" in exported_dict["contents"]["a"]["contents"]
     assert "A1" not in exported_dict["contents"]["a"]["contents"]
     assert "b" not in exported_dict
+    assert "g" in exported_dict["contents"]
+    assert "A3" in exported_dict["contents"]["g"]["contents"]
+    assert "A4" not in exported_dict["contents"]["g"]["contents"]
     exported_dict["contents"]["a"]["contents"]["A2"]
+    exported_dict["contents"]["g"]["contents"]["A3"]
 
 
 def test_create_and_update_allowed(enter_username_password, context):
@@ -206,8 +314,13 @@ def test_create_and_update_allowed(enter_username_password, context):
     alice_client["c"]["x"].update_metadata(metadata={"added_key": 3})
     assert alice_client["c"]["x"].metadata["added_key"] == 3
 
+    alice_client["g"]["A3"].metadata
+    alice_client["g"]["A3"].update_metadata(metadata={"added_key": 9})
+    assert alice_client["g"]["A3"].metadata["added_key"] == 9
+
     # Create
     alice_client["c"].write_array([1, 2, 3])
+    alice_client["g"].write_array([4, 5, 6], metadata={"project": "projectA"})
     alice_client.logout()
 
 
@@ -233,8 +346,11 @@ def test_public_access(context):
     for key in ["a", "b", "c", "d", "e"]:
         assert key not in public_client
     public_client["f"].read()
+    public_client["g"]["r"].read()
     with pytest.raises(KeyError):
         public_client["a", "A1"]
+    with pytest.raises(KeyError):
+        public_client["g", "A3"]
 
 
 def test_service_principal_access(tmpdir):